refactor: moved flatbufferbuilder to separate class
- move flatbufferbuilder to its own class file and renamed to flatbuilder
